refactor: 페이지 layout 컴포넌트 layout 폴더로 이동
