--- src/components/BaseLayout/Logo.tsx
@@ -1,9 +1,0 @@
-import logo from '@/assets/logo.png';
-
-export const Logo = () => {
-  return (
-    <a href="/">
-      <img src={logo} width="86" alt="선물하기 로고" />
-    </a>
-  );
-};
