+++ src/pages/Login.tsx
@@ -1,12 +1,13 @@
-import logo from '@/assets/login-logo.svg';
+import loginLogo from '@/assets/login-logo.svg';
+import { LoginLayout } from '@/layouts/LoginLayout';
 
 import { LoginForm } from '@/components/LoginPage/LoginForm';
-import { LoginLayout } from '@/components/LoginPage/LoginLayout';
+import { Logo } from '@/components/Logo';
 
 export const LoginPage = () => {
   return (
     <LoginLayout>
-      <img src={logo} width="86" alt="카카오 로그인 로고" />
+      <Logo src={loginLogo} alt="카카오 로그인 로고" width="86" />
       <LoginForm />
     </LoginLayout>
   );
