+++ src/pages/Theme.tsx
@@ -1,8 +1,8 @@
 import { useParams } from 'react-router-dom';
 
+import BaseLayout from '@/layouts/BaseLayout';
 import { themeHeaderData, themeListData } from '@/mocks/mockData';
 
-import BaseLayout from '@/components/BaseLayout/BaseLayout';
 import { ThemeContent } from '@/components/ThemePage/ThemeContent';
 import { ThemeHeader } from '@/components/ThemePage/ThemeHeader';
